convert hookpoints for version > 3.0.8
